+++ sherpa/astro/io/__init__.py
@@ -57,18 +57,17 @@ from configparser import ConfigParser
 import importlib
 import logging
 import os
-import os.path
 import sys
 
 import numpy
 
-import sherpa.io
-from sherpa.utils.err import DataErr, IOErr
-from sherpa.utils import SherpaFloat
-from sherpa.data import Data2D, Data1D, BaseData, Data2DInt
+from sherpa import get_config
 from sherpa.astro.data import DataIMG, DataIMGInt, DataARF, DataRMF, DataPHA, DataRosatRMF
 from sherpa.astro.utils import reshape_2d_arrays
-from sherpa import get_config
+from sherpa.data import Data2D, Data1D, BaseData, Data2DInt
+import sherpa.io
+from sherpa.utils.err import DataErr, IOErr
+from sherpa.utils.numeric_types import SherpaFloat
 
 config = ConfigParser()
 config.read(get_config())
