+++ sherpa/astro/io/crates_backend.py
@@ -25,9 +25,11 @@ import numpy
 
 import pycrates
 
-from sherpa.utils.err import IOErr
-from sherpa.utils import SherpaInt, SherpaUInt, SherpaFloat, is_binary_file
 from sherpa.astro.utils import resp_init
+from sherpa.utils import is_binary_file
+from sherpa.utils.err import IOErr
+from sherpa.utils.numeric_types import SherpaInt, SherpaUInt, \
+    SherpaFloat
 
 
 warning = logging.getLogger(__name__).warning
