+++ sherpa/astro/io/pyfits_backend.py
@@ -47,9 +47,10 @@ from astropy.io import fits
 from astropy.io.fits.column import _VLF
 from astropy.table import Table
 
-from sherpa.utils.err import IOErr
-from sherpa.utils import SherpaInt, SherpaUInt, SherpaFloat
 import sherpa.utils
+from sherpa.utils.err import IOErr
+from sherpa.utils.numeric_types import SherpaInt, SherpaUInt, \
+    SherpaFloat
 from sherpa.io import get_ascii_data, write_arrays
