+++ sherpa/data.py
@@ -123,11 +123,13 @@ import numpy
 
 from sherpa.models.regrid import EvaluationSpace1D, IntegratedAxis, PointAxis
 from sherpa.utils.err import DataErr
-from sherpa.utils import SherpaFloat, NoNewAttributesAfterInit, \
+from sherpa.utils import NoNewAttributesAfterInit, \
     print_fields, create_expr, create_expr_integrated, calc_total_error, bool_cast, \
     filter_bins
 from sherpa.utils.parallel import parallel_map_funcs
 from sherpa.utils import formatting
+from sherpa.utils.numeric_types import SherpaFloat
+
 
 warning = logging.getLogger(__name__).warning
